<refactor> modify: 关系修改，子弹-->武器-->角色; 伤害系数从WeaponComponent统一控制改为WeaponEntity内独立控制，实现同一子弹不同武器不同伤害

+++ Assets/Scripts/Client/Bussiness/BattleBussiness/Repo/BattleArmorItemRepo.cs
@@ -28,7 +28,7 @@ namespace Game.Client.Bussiness.BattleBussiness.Repo
             return all.Remove(entity);
         }
 
-        public bool TryGet(ushort entityId, out BattleArmorItemEntity entity)
+        public bool TryGet(int entityId, out BattleArmorItemEntity entity)
         {
             entity = all.Find((entity) => entity.IDComponent.EntityID == entityId);
             Debug.Assert(entity != null, $"护甲ITEM {entityId} 不存在");
